+++ apps/cowswap-frontend/src/modules/tradeFormValidation/types.ts
@@ -79,6 +79,7 @@ export enum TradeFormValidation {
   CurrencyNotSet,
   InputAmountNotSet,
   RecipientInvalid,
+  RecipientNotSet,
   NetworkNotSupported,
   NetworkDeprecated,
   BrowserOffline,