+++ libs/iframe-transport/src/iframeRpcProvider/IframeRpcProviderBridge.ts
@@ -38,11 +38,11 @@ export class IframeRpcProviderBridge {
 
   /**
    * Creates an instance of IframeRpcProviderBridge.
-   * @param iframeWidow - The iFrame window that will post up general RPC messages and to which the IframeRpcProviderBridge will forward the RPC result.
+   * @param iframeWindow - The iFrame window that will post up general RPC messages and to which the IframeRpcProviderBridge will forward the RPC result.
    *  Also it will receive some special RPC events coming from the wallet, like connect/chainChanged,accountChanged
    */
   constructor(
-    private iframeWidow: Window,
+    private iframeWindow: Window,
     private iframeOrigin?: string,
   ) {}
 
@@ -81,6 +81,7 @@ export class IframeRpcProviderBridge {
       // Listen for messages coming to the main window (from the iFrame window)
       iframeRpcProviderTransport.listenToMessageFromWindow(
         window,
+        this.iframeWindow,
         IframeRpcProviderEvents.PROVIDER_RPC_REQUEST,
         this.processRpcCallFromWindow,
         this.iframeOrigin,
@@ -101,6 +102,7 @@ export class IframeRpcProviderBridge {
     // Listen for provider meta info request
     iframeRpcProviderTransport.listenToMessageFromWindow(
       window,
+      this.iframeWindow,
       IframeRpcProviderEvents.REQUEST_PROVIDER_META_INFO,
       this.processProviderMetaInfoRequest,
       this.iframeOrigin,
@@ -170,7 +172,7 @@ export class IframeRpcProviderBridge {
 
     // Send the provider meta info to the iFrame window
     iframeRpcProviderTransport.postMessageToWindow(
-      this.iframeWidow,
+      this.iframeWindow,
       IframeRpcProviderEvents.SEND_PROVIDER_META_INFO,
       { providerEip6963Info, providerWcMetadata },
       this.iframeOrigin,
@@ -179,7 +181,7 @@ export class IframeRpcProviderBridge {
 
   private onProviderEvent(event: string, params: unknown): void {
     iframeRpcProviderTransport.postMessageToWindow(
-      this.iframeWidow,
+      this.iframeWindow,
       IframeRpcProviderEvents.PROVIDER_ON_EVENT,
       { event, params },
       this.iframeOrigin,
@@ -193,7 +195,7 @@ export class IframeRpcProviderBridge {
   // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
   private forwardRpcResponseToIframe(params: ProviderRpcResponsePayload) {
     iframeRpcProviderTransport.postMessageToWindow(
-      this.iframeWidow,
+      this.iframeWindow,
       IframeRpcProviderEvents.PROVIDER_RPC_RESPONSE,
       params,
       this.iframeOrigin,